refactor: separate layout store
