+++ src/components/layout/dialogs/Search.jsx
@@ -25,7 +25,8 @@ import { useTranslation } from 'react-i18next'
 import { useLazyQuery, useQuery } from '@apollo/client'
 
 import NameTT from '@components/popups/common/NameTT'
-import { useMemory, useLayoutStore } from '@hooks/useMemory'
+import { useMemory } from '@hooks/useMemory'
+import { useLayoutStore } from '@hooks/useLayoutStore'
 import { useStorage } from '@hooks/useStorage'
 import Utility from '@services/Utility'
 import Query from '@services/Query'
