+++ src/components/layout/dialogs/filters/Advanced.jsx
@@ -5,7 +5,8 @@ import Grid2 from '@mui/material/Unstable_Grid2/Grid2'
 import { useTranslation } from 'react-i18next'
 
 import Utility from '@services/Utility'
-import { useMemory, useLayoutStore } from '@hooks/useMemory'
+import { useMemory } from '@hooks/useMemory'
+import { useLayoutStore } from '@hooks/useLayoutStore'
 import { useDeepStore, useStorage } from '@hooks/useStorage'
 import Header from '@components/layout/general/Header'
 import Footer from '@components/layout/general/Footer'
