+++ src/components/layout/dialogs/filters/MenuTile.jsx
@@ -6,7 +6,7 @@ import FormatSize from '@mui/icons-material/FormatSize'
 import Settings from '@mui/icons-material/Settings'
 import { Grid, IconButton, Typography } from '@mui/material'
 import Utility from '@services/Utility'
-import { useLayoutStore } from '@hooks/useMemory'
+import { useLayoutStore } from '@hooks/useLayoutStore'
 
 export default function MenuTile({ data, rowIndex, columnIndex, style }) {
   const [name, setName] = useState(true)
