+++ src/components/layout/dialogs/filters/SlotSelection.jsx
@@ -7,7 +7,8 @@ import DialogContent from '@mui/material/DialogContent'
 import IconButton from '@mui/material/IconButton'
 import Divider from '@mui/material/Divider'
 
-import { basicEqualFn, useLayoutStore, useMemory } from '@hooks/useMemory'
+import { basicEqualFn, useMemory } from '@hooks/useMemory'
+import { useLayoutStore } from '@hooks/useLayoutStore'
 import { useStorage, useDeepStore } from '@hooks/useStorage'
 import { Img } from '@components/layout/general/Img'
 import { DualBoolToggle } from '@components/layout/drawer/BoolToggle'
