+++ src/components/layout/dialogs/profile/GymBadges.jsx
@@ -6,7 +6,8 @@ import { useTranslation } from 'react-i18next'
 import { useQuery } from '@apollo/client'
 import { useMap } from 'react-leaflet'
 
-import { useLayoutStore, useMemory } from '@hooks/useMemory'
+import { useMemory } from '@hooks/useMemory'
+import { useLayoutStore } from '@hooks/useLayoutStore'
 import Query from '@services/Query'
 
 import { VirtualGrid } from '@components/layout/general/VirtualGrid'
