+++ src/components/layout/drawer/SelectorItem.jsx
@@ -8,7 +8,8 @@ import Box from '@mui/material/Box'
 import Typography from '@mui/material/Typography'
 
 import { useTranslateById } from '@hooks/useTranslateById'
-import { useLayoutStore, useMemory } from '@hooks/useMemory'
+import { useMemory } from '@hooks/useMemory'
+import { useLayoutStore } from '@hooks/useLayoutStore'
 import { useStorage, useDeepStore } from '@hooks/useStorage'
 import { Img } from '../general/Img'
 import { ColoredTile } from '../general/ColoredTile'
