+++ src/components/layout/general/Menu.jsx
@@ -7,7 +7,8 @@ import ExpandMoreIcon from '@mui/icons-material/ExpandMore'
 import Collapse from '@mui/material/Collapse'
 import IconButton from '@mui/material/IconButton'
 import Utility from '@services/Utility'
-import { useMemory, useLayoutStore } from '@hooks/useMemory'
+import { useMemory } from '@hooks/useMemory'
+import { useLayoutStore } from '@hooks/useLayoutStore'
 import { useStorage } from '@hooks/useStorage'
 import useFilter from '@hooks/useFilter'
 import Header from '@components/layout/general/Header'
